services/claims-ai-engine/pyproject.toml
[tool.poetry]
name = "claims-ai-engine"
version = "1.0.0"
description = "AI-powered claims processing engine for NPHIES Pro Digital Insurance Platform"
authors = ["BrainSAIT Digital Solutions"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.2"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
pandas = "^2.1.4"
numpy = "^1.25.2"
scikit-learn = "^1.3.2"
tensorflow = "^2.15.0"
redis = "^5.0.1"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
celery = "^5.3.4"
requests = "^2.31.0"
arabic-reshaper = "^3.0.0"
python-bidi = "^0.4.2"
cryptography = "^43.0.1"
pydantic-ai-slim = {extras = ["ag-ui", "openai"], version = "^1.0.8"}
python-multipart = "^0.0.20"
prometheus-fastapi-instrumentator = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
